+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - documentation(securityhub): security hub findings description with prescriptive guidance
 - feat(config): added deploymentTargets for awsConfig since configuration recorder cannot be turned off selectively
 - feat(guardduty): add deploymentTargets and autoEnableOrgMembers settings for guardduty
+- feat(networking): add tags property to interfaceEndpoints to tag private hosted zones for interface endpoints
 
 ### Changed